requirements.txt
streamlit
lightning-sdk
Pillow
dspy-ai